Disable reading all sample manifests due to memory issues.
The memory required to load a several thousand sample manifest is in the
~1.5GB range.  It happens that the first 100 manifests contain several
of this size which makes it impossible to list them without the process
crashing.  Whilst research is made into how to reduce the memory
footprint it's easy enough to turn this ability off as clients only need
to interact with their own manifests, not everyones.