Hopefully some performance improvements.
Made sure validates_uniqueness_of is guarded so that it isn't run when
we already know the name of a sample to be unique.  Also found that
validates_presence_of on a polymorphic association is better done on the
component columns.

Better way for walking the positions on a plate so that they can be
loaded in bulk and hopefully the well creation is a little quicker for
plate manifests.