Changing the state of plates.
To change the state of a plate the code actually changes the state of
the TransferRequest instances for the wells of the plate where the wells
are the target.  The actual behaviour of the state change is based on
the plate purpose, as the initial plates in the pulldown pipelines also
need to start the pulldown requests.