Support pooling of wells.
In the pulldown pipelines there are points where the wells on one plate
are pooled into a single well on another.  This pooling is governed by
the submission: wells that are submitted together are pooled together.

When we reach the end of a pulldown pipeline individual wells are
transfered into corresponding MX library tubes.  The wells are
transferred such that the pulldown library creation request is honoured.

And if we're modifying the state of the MX library tube then we need to
update the state of transfer requests, and pulldown library creation
requests in certain circumstances.