Improve the performance of plate creation & well importing.
It was taking ~5s to create a plate and import 96 wells into it.  It now
takes ~0.5s.  This is because it is easier to perform an import and then
a reload within a transaction than perform some random modification of
the information in order to be able to do it outside of a transaction.