TransferRequest should physically do the transfer on start.
When you start a TransferRequest instance it should behave as though the
physical transfers are being made: copying the sample but duplicating
the tags.

The only issue here is that there can be multiple transfer requests into
one destination and assets cannot hold more than one sample at the
moment.