Fixes based on client API interaction issues.
Firstly the children of a plate purpose are in the JSON as 'children'
but the I/O behaviour meant that the element was 'plate_purposes'.  So
had to alter the code when merging into a paged list of results.

The transfers that involve at least one plate need to check that the
source, and the destination if required, are present before checking
that the well positions are valid.