Aliquoting model (THIS COMMIT FAILS TESTS!)
The concept of an aliquot is fairly simple: it's an amount of stuff (at
the moment a pair of sample and an optional tag) that exists within a
receptacle.  Receptacles can contain multiple aliquots, which eliminates
the need for asset graph hacks for multiplexing and makes things a
little more consistent.

The introduction of aliquots means that TagInstance is retired, which
has a knock-on affect on the warehouse which has yet to be addressed.

As part of the aliquoting model the TagInstance class is now dead.  This
commit is all about doing that.  There are some changes that are worth
noting:

- app/views/batches/show.xml.builder has been altered slightly to remove
  references to the library ID, which will be gone as that is part of a
  hack for the TagInstance model.

- the migrations have been thoroughly rewritten to be TagInstance
  agnostic, but also to walk the asset & request graphs to build the
  aliquots.

The migration to map the samples and tags to aliquots will will walk the
asset graph and skip to the request graph for pulldown multiplex library
creation requests, which happen not to establish a parent-child
relationship.

There have also been some behavioural changes:

TransferRequest now has to transfer the aliquots from the source to the
destination at creation time as empty wells cannot be tagged.  Before it
was doing this on state change which, although physically correct, isn't
going to work (at the moment).

There is only one place in this code that is now failing the tests:

The ordering of the wells on a plate seems to have broken sample
manifests / volume checking.  Completely unrelated to aliquots.