Refactored API authorisation to use CanCan.
By using CanCan the control of authorisation has been made much simpler
and far more flexible.  The behaviour remains the same, for the moment,
but can be changed to be considerably more powerful without needing
massive code changes.

Managed to work back in the feature that tests the behaviour of the API
from just the basic functionality perspective.