[fixes 14691821] Plug a potential memory leak in the API GC handling.
If an action raised an exception then the GC wasn't re-enabled.
Although this is unlikely to be happening in production the cucumber
features do this all of the time.  Before this change the code ran the
risk that the time between an exception from the API to the point where
a valid request was made, could be a while and so the memory could
become used up.