Performance enhancements.
Rather than having to set all of the requests leading from the stock
plate wells, we only need to set those that are not transfer requests,
for initial pulldown plates.

To calculate the iteration of a plate we only need to look at the parent
asset links to other plates of the same plate purpose.  The trick is to
get MySQL to do the row number so that we don't have to.

Tidied up the acts_as_descriptable code because it was a mess.  Still
needs a lot of work.