Addition of searching for pulldown plates & some code tidy ups.
This adds the ability to find all of the plates that are in the pulldown
pipline that have not been transfered from.  This covers all plates that
are both waiting for work (like a stock plate might be) and are in
progress.

Did some minor code tidy ups, particularly surrounding the use of
class_eval which, in places, didn't include the file and line number.