Fixes some issues with user assignment & DB ordering.
The DB seems to make up its own mind when it wants to listen to a query
that has order and limit in it, so we force it by doing one query with
order and then limiting that with another.

Missed off a load of user attribute mappings for the various things that
need it.