Ensure that the request behaviour is similar to production.
When testing submissions make sure that the wells on the plate are
submitted in the same manner as the user would, i.e. column major order.
Do not rely on the order of plate.wells because that wasn't guaranteed
and is ordered now.

Use the appropriate request ordering behaviour in pipeline.