Expose pools for plates & ensure duplicate tags handled earlier.
This should ensure that if someone tries to tag a plate with tags that
would cause duplicates later (based on submission) that they are failed
at that point, rather than getting to the end of the pulldown pipeline.

We need to expose pools on the plate so that the pulldown application
can reduce the set of tag templates that can be used.