Requests are being tricky with when aliquots should be copied
In the migration to aliquots we simply follow all requests and blindly
assume that this is the correct thing to do.  However, in the code we
only update aliquots when they the request moves to started.

We also need the request to belong_to receptacles, rather than assets,
so that the aliquots association is correctly available.  This does mean
that this code will not work in development mode, where STI isn't
properly handled in Rails auto-reloading.

Ensure that the SpikedBuffer aliquots are also properly updated.