Various fixes from testing in the next release environment.
The progress of anything is 0% if there are no passed requests.

Various things need UUID, like item and tag group.  And the qc_state on
a batch comes from qc_state, not compatible_qc_state which is on asset.

Updated the bait library information in the seeds and fixed the fragment
sizes in the migration and the seeds as the column is a string.

Added a batch limit to the cluster formation pipelines as these had nil
or 2, both of which are wrong!