Couple of bugs & performance enhancement.
Controls were being added in a funny way to batches meaning that they
ended up in completely the wrong position and upset the ordering.

Requests that were not ready were showing up in the pipeline inboxes
causing genotyping to display a lot of red warnings.

Assets for a sample should be eager loaded to drastically improve the
performance of the sample show page.