Performance enhancement for finding the stock plate.
This code uses a much quicker way of finding the stock plate from a
given plate.  It basically counts the number of plates between the
specified plate and a stock plate for one well, as all wells on the
specified plate come from a stock plate at the same depth (though not
necessarily the same stock plate).  It then does enough joins of the
requests table to itself to find the appropriate stock wells.

This is ~15% quicker than walking back for each well over a short
distance, and will only be more efficient over longer ones.