[fixes 18141977] Tagging obeys plates then pools.
The tagging will ensure that tags are laid out in the correct order
regardless of pools.  So each row will have the same tag across its
entire length in the case where there are no long pools.  In cases where
the pool is longer than the column extra tags will be used to avoid
clashes but as substitutions for the overlap.