[fixes 19151925] Use SQL to identify pools for a plate.
Identifying pooling for a plate can be made by querying the DB, rather
than determining it using code.

This code also modifies the Map model so that ordering by row or column
can be done at the database level.  It removes the explicit ordering on
plate.wells and leaves it to the caller to decide the order.

Discovered that at least one feature had a different ordering to all of
the others in one of its tables.  So made it consistent with the
majority.