[fixes 19232749] Layout tags correct with empty wells.
This code makes the tag layout code behave as though a tag is placed in
an empty well, effectively skipping over the tag.  This behaves as
expected for 96 tags but may behave oddly in the presence of complex
layouts as it pretends that any empty wells are part of the pool prior
to them in the group.  In other words, if you were going column wise
then B1 would be considered in the same pool as A1; if you were going
row wise then A2 would be considered in the sample pool as A1.