[fixes 19887963] Fixes for the various tests
Removed a scenario that was on about putting a request in two batches
but never actually did and was actually testing something that is tested
everywhere else that batches are used.

Batches created for the gel feature were not correctly setup, so changed
the step that was used and removed the broken one.

Spiked in controls were failing because of the unpredicatability of the
sample ID, so ignored it in the XML check.