[fixes 19881309] Name for label should be implementation specific.
The 'study' attribute for a barcode label comes from different places
depending upon the asset.  For tubes it comes from the primary aliquot
sample ID, or the name of the asset if that is not present.  For plates
it is the name of the plate.

All other assets are unprintable and therefore should error.