[fixes 19943167] Submission request options not behaving from API.
Submissions created through the API do not behave properly if the
request options are set.  They pass the submission level validation but
then fail at the request graph generation.  There are also problems
where the default options from the template do not get maintained and
are actually seen as different when they do appear.

This ensures that all of the numeric fields are treated as such, and
that API submissions get their options properly merged into the
defaults.