[fixes 19890039] Hide batched pending requests from inbox.
Once requests are batched they used to disappear from the inbox as the
batch would automatically be started.  Now, however, that does not
happen so we need to hide them based on whether they have been batched
or not.

This includes a bit of a code tidy in the request class, moving the
batch related behaviour into an appropriate module, and it removes a
couple of unused methods.