[fixes 18272943] Do not allow untagged batches out of MX pipeline
Batches that have not been tagged should not be allowed to be completed
in the MX library creation pipeline.  So this code adds a final
validation when the batch is in that state to check.

In the process I discovered that the MX library creation pipeline in
the seeds file was of the wrong type, which lead to a bug in the batch
reset feature!