[fixes 15313119] Fix priority flag issues.
This fixes the issue where the priority flags for MX library creation
requests wasn't working.  It refactors the code so that it goes through
one action and uses one set of Javascript.

It also tidies up the priority ordering of the table so that an
additional column isn't needed, although this appears to have required
the uncompressed version of the tablesorter javascript.