[fixes 20472069] Workflow stage 'cancel' should return to batch
Makes the 'cancel' link return to the batch, rather than the pipeline
inbox.

Tidies up the views so that they call the same partial to render the
submit and cancel controls, and can uniformly control the form
submission behaviour.

Reworked the cherrypicking Javascript a bit to make is somewhat more
readable, but there's a lot of work needed on it.