[fixes 20927351] Enabling orders & submissions through new API.
This enables the new API to support orders and submissions as
independent resources, allowing clients to create orders and add them to
a submission.

Renamed the SubmissionTemplate endpoint to OrderTemplate as this is
really what it is, but had to alias the former to the latter so that the
lookup still works for the instances.

Fixed a bug in the API that allowed registered users to create asset
audits, which is not what we want.