[fixes 20927351] Use order creation and submission addition.
Rather than using the "create a submission with these details" step I've
split this into "create an order with these details", "add the order to
a submission" steps.

In the process discovered that some of the features were missing
resources with the UUIDs that were being used, so corrected them.