[fixes 21584975] Only check required options if specified
The order now only checks the required options if they are specified or
when it is being built.  This means that if someone does not specify
read_length at construction then the order will not throw an exception,
but it will when they try to submit the submission it is in.