[fixes 21906265] Remove unpooled wells & rollback last change
Basically all wells need to be present with their pool ID, even though
that may be nil, for the tag layout code to work appropriately.
Otherwise the offset of the tag used is incorrect.

The easier way to solve the pulldown issue is to not return an
"unpooled" pool and simply ignore those wells and deal with it at the
pulldown application.