[fixes 23142589] fix submission#show exception
The exception occurs an order is saved with a list of samples.
The submission is without reloading the order, therefore the
Order#complete_building method in charge of creating the asset group is
not called.

We force the reload of orders in complete_building

All test passes but some features using the old submission GUI fail.