[fixes #2457273] Fix up symbol/string mismatches in requests.
Ensure that the attribute names in the submission templates match those
from form parameters, i.e. they are strings.  This then prevents the
duplication of the values when people enter something different.

Then we need to ensure that if attributable classes override behaviour
that they must force it to happen, this ensures that the fixed fragment
size information for pulldown requests comes in correctly.