[fixes #25247387] Handle request graph changing for submission.
Sometimes a submission, particularly in the case of library creation,
can go through pipelines in several batches.  If these batches then
cause requests to be created (for example, failing a sequencing batch
does) then the request graph for a submission can change and the
behaviour of Submission#next_request is not guaranteed.

This commit changes that so that it works with the original part of the
request graph based on the orders of a submission.