[#27906689] Some changes to improve performance
Don't need to validate the uniqueness of content_id in
container_associations as that's enforced by the index on the table.

Fixed a problem with the cherrypick plate generation where the array
needed to be duplicated, before being cleared.

Use study_id & project_id when dealing with the aliquot cloning to
prevent a query of the DB.