[#27906689] General improvements in record creation
Removed a load of validations that are already being handled by the
constraints in the database.

Create UUID directly and assign to the record, rather than creating
through the record, which causes a SELECT query to be performed.

Wrap certain operations within a transaction to improve the performance
of AMQP broadcast, as well as protect the data.