[fixes #28830149] Ensure correct updated_at in JSON.
If an association is included in the JSON data for the warehouse
builder, then the updated_at should be the most recent date across all
associations.  For example, the JSON for library tubes depends on the
primary aliquot and, if that changes and the tube doesn't, then the
updated_at for the tube should be the updated_at for the primary
aliquot.

Essentially the JSON is a view of an entire object and the updated_at
therefore encompasses all minor components of that object.