[fixes #28937311] All for unsaved records to have no UUID
If a record does not have a UUID then one is created for it;
unfortunately this doesn't work when the record is unsaved, as the ID
information is nil and that violates a DB constraint.  This commit
adds the ability for the code to say "this record should not have a
UUID", and then that will not cause errors, except if they try to save
this marked record (which is an error in itself).  Hence the integrity
of UUIDs should be maintained and the ability to preview stuff (like
bait library assignment and transfers) is supported.