[#30132373] Refactored cherrypick code to be more readable
Unit test for cherrypick is considerably better and the code is slightly
easier to read as you pick into a PickTarget, which puts the behaviour
of wells in there, i.e. it was hard to spot how the empty wells, control
well, and template wells all went into the pick.