[#32452847 #32455269] Added tube purposes for Illumina B
Illumina B has stock MX library tubes and MX library tubes.  These
behave in the same manner as Illumina A, which is to say, not the same
as the legacy system.

Added a request type for the transfers between the PCR XP plate and the
stock MX library tubes for Illumina B.