[#30918567] Cherrypick for Illumina now with optimum strategy
This refactors the cherrypick for Illumina code such that it will use a
pick strategy required by the plate purpose.  In the case of Illumina B
this means an optimum packing strategy, where empty space is filled by
plexes that reduce it.  Also, it ensures that plexes are not split over
one of the dimensions (width in the case of Illumina B) if they can be
placed within a dimension on their own.