[#37925787] Refactor cherrypick to be filter based
This refactors the cherrypicking strategy so that it is filter based:
plexes are passed through a sequence of filters that reduce, order or
manipulate them such that the next appropriate plex is first in the
list.  Each plate purpose now describes the ordered list of filters to
be used during a pick.