[#38415325] Dummy paginate has_many associations
When generating the actions that are available for a has_many
relationship the code uses the will_paginate 'paginate' method.
Unfortunately this is not lazy, loading up to per_page objects into
memory, which may then not be used.  This commit provides an object that
behaves like a will_paginate collection but does not load any of the
physical objects.