[#40535747] Hack Net::HTTP::ProxyDelta for our proxy
This initializer solves the issue seen with our corporate proxy
rejecting CONNECT requests by allowing the code to set additional
headers for the request.  It then sets the one header that is required
to stop rejection: User-Agent.

This is a nasty, nasty hack because it's a cut-n-paste of the connect
method from the JRuby 1.6.8 code, so if anything changes then this needs
to be updated.