[#53289573] Start batch task starts batch (and requests)
Previously all tasks would start the batch on render.
This adds a new task type, which will allow us to
decouple the behaviour, ensuring the starting of
batches reflects the start of labwork.

This task is designed to be transparent.

Note: This commit only adds the tasks. They are
implimented in the next commit.