Control requests have own methods
In batch XML the decision to display like a control would be handed
off to the asset. This resulted in some flowcells having multiple controls.
This behavior was always at odds with the way analysis worked, but the
discrepency only affects 20 batches. Although this might cause some
confusion if anyone ever looks back at old batches, it was decided
that thew new behavior is 'correct'.