[#84299082] Expose request type key in pool info
We expose the request type key to help use determine
what plate is needed next. Pool and Pippin use
target_tube_purpose which isn't available to us here
as we have no target tube.

We use key rather than name, as despite the
inconsistency with a similar technique in the Generic
Lims, it is a more robust choice. Uuid would be even
more robust, but is not human readable.