[#85886858] Fallback to empty hash for creation_options in the event of missing parent
I'm not certain about this fix, as it is a bit of a patch on bad data.
I'm going to be fixing the actual problem underlying out current issues,
which is the wrong purpose being assigned to MX library tubes from certain pipelines.

This fix does make things more pleasent for the users, but could result in more
serious underlying issues getting overlooked.