Use request.id ASC as fallback sort for batch.requests
While strictly speaking we never specified a batch sort
order before, in practice MySQL would almost always
retunr the requests in id order, and some code assumed this
would be the case. When we specified position however
MySQL tended to flip the order requests were returned in
when position was NULL. This causes issues with fluidigm
where plate order is assumed.