[#95420840] Rename index_tag to tag_2
After discussion with NPG we have renamed the
index_tag to tag_2 to align ourselves with the
approach of Illumina, and to better allow for
future extension.

Note: If we do end up with a third tag at any point
we may be better off pulling the aliquot-tag
association out into a join table with a tag_index
column. This has consequences wrt. to maintaining
tag uniqueness though.