Strip out empty plates and rase cherrypick errors for invalid picks
Previously running onto two plates would result in a second plate
creation, even if the previous was empty. This was especially
a problem for cases where a fluidigm barcod ewas assigned. Here
we also detect cases where multiple plates are going to be created
with the same barcode, and raise a cherrypick error, which gets
handlede more elegantly by the controller.