Re-run copyright date tagging script
While updating to rails 3/ruby 1.9+ it was necessary
to make changes that affected many files at once.
Manually updating the timestampls would be a pain, not to
mention the cause of multiple merge conflicts. So I left updating them
until everything was good to go. Unfortunately this also means that
we seem to have applied a timestamp for the initial application of
the copyright message. Ideally we'd filter out all such insubstantial
commits.