Add find plates for user
In attempt to curb the rapidly expanding number of subclasses for
searches we add parameters to the search table, allowing us to
define default options, and ultimately expose more complicated queries.

We're just introducing a proof of concept here, but we'll follow
it up with shifitng the existing templates to the new approach.