Exchange is passed into active exchange, not set on an instance variable
Observers are singletons, so this behaviour was not thread safe.
While a new client was created each time, the resulting
exchange was saved on an instance variable, and thus shared between
concurrent blocks. This meant that when one client closed at the end
of the block, it would disrupt the excahneg in concurrent blocks.

Instead we pass the exchange into the block, removing state from the process.