Push recused exceptions onto @bulk_submission instead of onto hash
Clearing the flash caused a large number of test failures as the flash
was no longer accessible to the test suite. This might explain why it
is persisted until the start of the next action.

Instead we write a more generic message t the flash, and push the errors
onto base. We also tell formatastic to render any errors on base.

In practice we probably should be handling these errors before then
raise an exception.