Source volume is passed in directly
Previously source volume was mistakenly pulled from the well_attributes
which actually reffered to the target well. Fortunately this was nil,
so got compacted out, rather than resulting in a zero pick.

Here we instead pass in the source volume directly, and use that instead.
We pass in the volume as an extra argument, rather than just sending the
source well itself to reduce coupling.