Hotfix - Stock metric found for all stock wells
The attempt to look up stock metrics was raising an exception
for re-sequencing requests from pools, as these had too many
stock wells. However, this is entirely legitimate behaviour,
the raising of an exception was over cautious.

Instead we map over each of the stock wells and find the latest report.
This is consistent with the behaviour on building of the original
submission.