Rename ERA to ENA
The archive is called the ENA. The ERA is a historical name, which has not
been used for a while. To avoid confusion, I've renamed it here.

Note: The configuration in the deployment project will need to be update.
I'll make a commit shortly after this with both versions to avoid us
accidently forgetting.