Paginate results and reduce memory consumption
Large studies were causing high memory usage when viewing these summary
pages. The pagination helps to prevent this being a major issue
and some improvements in loading our assets should also help.

This controller needs a substantial re-visit. The code as is
could do with refactoring, but also the general contents of the
page are not very user-friendly. Not only that, but the statistics
have been misleading since cross study pooling.