[ss526] Allow library manifests to update sample metadata
Note: This doesn't support the updating of library information as
it is not possible to do so on a per row basis due to multiplexing
and tag uniqueness constraints. It would be neccessary to add some
kind of callback on manifest completion. This behaviour MAY get
added at a later date.