Pipelines define the required batch sheet directly
The previous mechanism was confusing and error prone, using
properties on batch (most of which delegate to pipeline)
that selected the correct template through correlation
rather than any direct association between the property
and the requirement. Here it becomes a propery of the pipeline.

This still needs refinement. We can probably look at using the
rails template system a little more directly.