Rescue ActiveRecord::StatementInvalid when processing manifests
We had a few cases where statement invalid was being raised in
the event of some character encoding issues. Delayed Job would then
attempt to record the exception in the database, fail (due to the
invalid characters) and then die. While we need to properly fix
the handling of character encoding, this change addresses the immediate
production concerns.

A monkey patch of delayed job would fix the issue globally, an
example is seen in the issue tracker:
https://github.com/collectiveidea/delayed_job/issues/774