ss586 - QC tubes don't remove aliquots on retrospective fails
MiSeqQC is used to determine if samples are sufficient quality
before progressing to pooling. In the event they are at low
concentration they may be failed on the plate from which the
MiSeqQC tube was made. Prior to these changes, this failure
would trigger te removal of downstream aliquots, including
those in the MiSeqQC tube. This was undesireable.

This disables aliquot removal in QC tubes.