Simplify iteration calculations
The interation calculation was previously a complicated custom
SQL query. Here is is updated to use standard rails methods.
This change was prompted by odd brittle behaviour, resulting from
some plates not finding themselves, and the select_one query returning
nil. While this doesn't fix the underlying issue, the code is more
robust, and easier to read.

As best I can tell, the underlying issue must be a result of a
race condition, in which the plate purpose changes between being
loaded, and the query being performed. This is quite worrying.