Remove Sample Manifest core test
Sample manifest generation is slow and this test takes 30s to run as
a result. Meanwhile, this code is currently only used for RapidCore
manifests in production. While testing this behaviour is probably
'correct' I've removed the tests for performance reasons. We can
add them back in if we start relying on the behaviour, or if
manifest generation improves.