Prevent release og Active Record connection on api close
Rails connections are usually handled as part of the request
lifecycle, as part of the rack stack. This includes actions
such as cleaning the rails query cache. The premature release
of the connection to the pool here was resulting in the
query cache persisting over multiple requests to the API.