Use round rather than string interpolation for float conversion
In MRI
"%.1" % 6.25 # 6.2
6.25.round(1) # 6.3

In Jruby
"%.1" % 6.25 # 6.3
6.25.round(1) # 6.3

While round to even is a *better* solution the inconsistency
was making tests tricky, and we aren't using these figures
for statistical processing, so don't care especially about the
bias.

It looks like BigDecimal does provide support for rounding to even
if we do decide to go that way.