seq355 - Remove unused and non-functional size check
We have no pipelines with this attribute set, and
the code looks non-functional. The only test covering
it was in a file flagged wit, so wasn't being run.

While the ability to limit the number of plates
in a batch IS of interest to cherrypicking, the
main interest is in limiting the number of source
plates for TECAN file generation. There are better
ways of handling this.