Exclude wells in other submissions
Wells in other submissions were still included in the each
and would fail when trying to find an appropriate outer request.
This was causing issues if all wells in a submission were failed
as they would no longer be listed as pools in active limber.
It would also cause issues if we ever did end up passing
partial plates.