Speed improvements for large transfer request collections
- Switch order of trarget tube and transfer request loading

Specifying the loading of target tubes AFTER the loading of
transfer requests seems to unload the transfer requests, probably
because of the has_many through realtionship on target tubes.

- Manually extract uuid resources for bulk create.

The old API retrtieves UUIDs in a very inefficient manner, as
each resolved template looks up its uuids independently. This
can be quite flexible, but is impossibly slow for large bulk
createds of nested records.