SEQ622 - Add outer request to aliquot
This is an initial experimental spike to look at the
feasibility of persisting outer request on aliquot.
This is to improve performance and reliability of the final
library passing.

Generally it was pretty easy to apply the outer request
given the new transfer request behaviour. THis made lookup
at the other end faster, with performance tests droping from
120s to 80s. However there were still substantial performance
issues, mainly associated with the updating of aliquots.

Changes were made to work completions to lower the update
time to ~20s. This is still slower than ideal, but further significant
improvements will be tricky at this stage.