SEQ654 - Fix failure to assign billing product via linear submission
release-10.0.0 included changes to the request constructor to remove
some meta-programming. These included a modification to the way in
which billing products were assigned in the request factory. These
modifications set-up the billing product in the create block, rather
than on a subsequent update after creation, avoiding the need to hit
the databaser twice. However, while the tests covered the situation
in which metadata was set via the hash passed in at initialization,
they did not cover any attempts to set metadata within the create
block. This process was used in submission building.

- Add tests for assigning billing products when metadata has been
 set via a block passed to create.
- Re-ordered asignment of billing product until after any custom
 blocks have returned.