Simplify PlateType lookup
- Remove fallback to external properties
- Remove magic string handling for 'Stock Plate' type
- Rename getters to follow Ruby convention
The old code used to fallback to the external properties, and would
also have special handling of 'Stock Plate' types.
External properties were previously imported from the Oracle Cache,
which is no longer used. In addition, a quick trawl of the production
database indicates that plate types were NEVER imported, as a result
the 'Stock Plate' ype never existed.
As a result, its been possible to drastically simplify the code.

Following on from this it might be worth switching things to a
simple association, rather than the odd serialization.