NPG requests are idepotent
If NPG believe a request has failed, they will re-attempt it.
Occasionally this happens despite us recording the request as
passed. Previously we would reject the request if it was a duplicate,
now wer only reject requests which differ. Duplicate requests
just trigger the render (no new event is recorded)