If a manifest fails to process, flag as failed
Currently failed manifests get stuck at processing, leading
users to believe that they will process if they wait long enough.
This change flags them as failed instead.

That said, this isn't a complete fix, as it doesn't help
indicate what has gone wrong, and realy failure should
be caught at the earlier validation anyway. It may be more
useful if we throw an exception here as well, as well as
logging eaxtly which bit of the completion check failed.