Reset the uploaded document association before failing
When processing the manifest we update the uploaded document
with the new file. If the manifest fails, the transaction gets
rolled back, but the new document is still flagged for persistance.
When the manifest if fail!ed this triggers an autosave of the
newly created uploaded file, which then promptly fails, because
Rails doesn't set documentable_type. Not quite sure why the rollback
breaks the association, but we don't want this persisted anyway.