Identify requests directly via submission
Popping via stock plate makes things brittle is additional stock plates
are introduced. This change reduces the risk of introducing tag
clashes, although does mean that we'll detect cross-plate submissions
on the basis of the initial requests in the submission only. This
is currently not a problem, as we don't have any submission templates
active which put cherrypicking upstream of library creation. And it
seems likely that the point will be rendered moot by the time that
changes.