Fix finding next requests for mixed submissions
The problem:
Finding next requests is surprisingly convoluted, and uses the
offset of the request within the the submission to match up
with the equivalent offest request for the next request type.

Unfortunatley these offsets get mixed up if the two orders
have a mixture of shared and unique request types prior to
multiplexing.

This change takes the order into account.

next_requests still needs some clean up, but at this point tests
are passing.